refactor(core): rename http_request.dart to api_client.dart + clean up ApiClient
- Rename file to match class name: http_request.dart → api_client.dart
- Remove code duplication with _execute() helper method
- Add dart doc comments on all public methods
- Remove section separator comments
- Use latest Dart features (switch expressions, sealed class)
- Update CODE_STYLE.md guidelines: file name must match class name
