+++ lib/ui/viewmodels/animes/anime_search_controller.dart
@@ -6,7 +6,7 @@ import 'package:application/data/models/anime_dto.dart';
 import 'package:application/data/models/pageable_dto.dart';
 import 'package:application/data/models/enums/search_type.dart';
 import 'package:application/core/analytics/analytics.dart';
-import 'package:application/core/network/http_request.dart';
+import 'package:application/core/network/api_client.dart';
 import 'package:application/core/network/api_result.dart';
 import 'package:application/core/widgets/widget_builder.dart' as wb;
