+++ lib/ui/viewmodels/animes/missed_anime_controller.dart
@@ -3,7 +3,7 @@ import 'package:application/ui/viewmodels/member_controller.dart';
 import 'package:application/data/models/missed_anime_dto.dart';
 import 'package:application/data/models/pageable_dto.dart';
 import 'package:application/core/constants/constant.dart';
-import 'package:application/core/network/http_request.dart';
+import 'package:application/core/network/api_client.dart';
 import 'package:application/core/network/api_result.dart';
 import 'package:application/core/widgets/widget_builder.dart' as wb;
