+++ lib/ui/viewmodels/episodes/followed_episode_controller.dart
@@ -2,7 +2,7 @@ import 'package:application/ui/viewmodels/generic_controller.dart';
 import 'package:application/ui/viewmodels/member_controller.dart';
 import 'package:application/data/models/episode_mapping_dto.dart';
 import 'package:application/data/models/pageable_dto.dart';
-import 'package:application/core/network/http_request.dart';
+import 'package:application/core/network/api_client.dart';
 import 'package:application/core/network/api_result.dart';
 import 'package:application/core/widgets/widget_builder.dart' as wb;
