+++ lib/ui/views/anime_details_view.dart
@@ -22,7 +22,7 @@ import 'package:application/data/models/season_dto.dart';
 import 'package:application/l10n/app_localizations.dart';
 import 'package:application/core/analytics/analytics.dart';
 import 'package:application/core/constants/constant.dart';
-import 'package:application/core/network/http_request.dart';
+import 'package:application/core/network/api_client.dart';
 import 'package:application/core/widgets/widget_builder.dart' as wb;
 import 'package:flutter/material.dart';
 import 'package:share_plus/share_plus.dart';
